refactor dpm campaign assignment task routes
